refactor: Moving user and workspace to v0 schemas
